telegram-115bot/requirements.txt
python-telegram-bot==20.7
qrcode==7.4.2
pillow==10.1.0
pyyaml==6.0.1
requests==2.31.0